datos correctos para members desde general_meeting + mejora formulario

+++ app/controllers/members_controller.rb
@@ -1,11 +1,12 @@
 class MembersController < ApplicationController
     #GET /members
     def index
-        @members = Member.all.order('id DESC')
+        @members= Member.where(members: { general_meeting_id: params[:general_meeting_id] })
     end
 
     #GET /members/new
     def new
+        @general_meeting_id = params[:general_meeting_id]
         @member = Member.new
     end
 
@@ -14,7 +15,7 @@ class MembersController < ApplicationController
         @member = Member.new(member_params);
         if @member.save
             flash[:notice] = "Miembro creado con éxito."
-            redirect_to :action => 'index'
+            redirect_to :action => 'index', :general_meeting_id=>params[:member][:general_meeting_id]
         else
             render :new
         end
@@ -28,26 +29,27 @@ class MembersController < ApplicationController
     # PUT /members
     def update
         member = Member.find(params[:member][:id])
-        member.update_attributes(member_params) 
+        member.update_attributes(member_params)
         flash[:notice] = "Miembro ha sido editado con éxito."
-        redirect_to :action => 'index'
+        redirect_to :action => 'index', :general_meeting_id=>params[:member][:general_meeting_id]
     end
 
     # DELETE /members
     def destroy
         member = Member.find(params[:id])
-        #if Member.where(member_id: member).empty?
+        if MeetingMember.where(member_id: member).empty?
             member.destroy
             flash[:notice] = "Miembro eliminado con éxito."
-            redirect_to :action => 'index'
-        #else
-        #    flash[:notice] = "el miembro a eliminar tiene elementos asociados."
-        #    redirect_to :action => 'index'
-        #end
+            redirect_to :action => 'index', :general_meeting_id=>params[:general_meeting_id]
+        else
+            flash[:notice] = "el miembro a eliminar tiene elementos asociados."
+            redirect_to :action => 'index', :general_meeting_id=>params[:general_meeting_id]
+        end
     end
 
     private
     def member_params
-      params.require(:member).permit(:primer_nombre, :segundo_nombre, :primer_apellido, :segundo_apellido, :dui)
+      params.require(:member).permit(:primer_nombre, :segundo_nombre, :primer_apellido, :segundo_apellido, :dui,
+      :charge_id, :general_meeting_id, :activo)
     end
 end
